refactor: adding responsivity
