+++ app/components/VideoBackground.tsx
@@ -7,9 +7,9 @@ import Link from 'next/link';
 export default function VideoBackground() {
     return(
         <section className="w-full h-full flex flex-col items-center justify-center relative mb-8">
-            <div className="absolute w-full top-0 flex flex-col justify-end p-8 items-end text-center">
-                <div className='flex flex-col items-center z-10'>
-                    <p className="p-4">feel free to explore!</p>
+            <div className="absolute w-full top-0 flex flex-col justify-end xl:p-8 md:p-6 items-end text-center">
+                <div className='flex flex-col items-center z-10 md:mt-6 gap-2'>
+                    <p className="xl:text-xl">feel free to explore!</p>
                     <Link href="#explore" className='cursor-pointer'>
                         <ArrowDown className="animate-bounce" />
                     </Link>
